Added generic class + all the files with renamed directory
