+++ Typeracer/Services/IGraphService.cs
